🐛 Move `torch.compile` comptability check function to avoid `test_logger_output` error
