+++ tests/test_init.py
@@ -154,6 +154,6 @@ def test_lazy_import_module_not_found() -> None:
 
 def test_torch_compile_compatibility() -> None:
     """Test if torch-compile compatibility is checked correctly."""
-    from tiatoolbox import is_torch_compile_compatible
+    from tiatoolbox.models.architecture.utils import is_torch_compile_compatible
 
     assert isinstance(is_torch_compile_compatible(), bool)
